backend/pyproject.toml
[tool.poetry]
name = "qode-clarity-backend"
version = "0.1.0"
description = ""
authors = ["IGWE, Chibueze"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
firebase-admin = "^6.7.0"
google-cloud-firestore = "^2.20.1"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
pydantic = "^2.10.6"
starlette = "^0.46.1"
google-cloud-storage = "^3.1.0"
google-api-python-client = "^2.165.0"
python-dotenv = "^1.1.0"
python-multipart = "^0.0.20"
email-validator = "^2.2.0"
httpx = "^0.28.1"
markdownify = "^1.1.0"
aiofiles = "^24.1.0"
markdown = "^3.7"
beautifulsoup4 = "^4.13.3"
jinja2 = "^3.1.6"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
